Created joystick control for lift

+++ src/org/usfirst/frc/team5987/robot/subsystems/LiftSubsystem.java
@@ -47,7 +47,7 @@ public class LiftSubsystem extends Subsystem {
 		return clawMotor.get();
 	}
 
-	public void setClawSpeed(int speed) {
+	public void setClawSpeed(double speed) {
 		clawMotor.set(speed);
 	}
